refactor: move EventNames enum to separate file
