+++ packages/realtime-api/src/client.ts
@@ -12,7 +12,8 @@ import VERTC, {
 } from '@volcengine/rtc';
 
 import { getAudioDevices, isScreenShareDevice } from './utils';
-import { EventNames, RealtimeEventHandler } from './event-handler';
+import { EventNames } from './event-names';
+import { RealtimeEventHandler } from './event-handler';
 import { RealtimeAPIError, RealtimeError } from './error';
 import { type VideoConfig } from '.';
